⚡️ bench: Use F64 instead of F32

+++ packages/bench/suites/n-body/components/Circle.ts
@@ -1,5 +1,5 @@
 import { Types, defineComponent } from "@bitecs/classic";
 
 export const Circle = defineComponent({
-  radius: Types.f32,
+  radius: Types.f64,
 });
